refactoring: moved tutor into his pkg
