+++ src/main/java/com/example/primo_progetto_spring/Entity/Classroom.java
@@ -2,6 +2,7 @@ package com.example.primo_progetto_spring.Entity;
 
 import com.example.primo_progetto_spring.Student.Studente;
 import com.example.primo_progetto_spring.classprogram.ClassProgram;
+import com.example.primo_progetto_spring.Tutor.entity.Tutor;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.*;
